+++ crypto/sessionkey_streaming_test.go
@@ -257,7 +257,7 @@ func TestSessionKey_DecryptStreamCompatible(t *testing.T) {
 		&PlainMessage{
 			Data:     messageBytes,
 			TextType: !testMeta.IsBinary,
-			Time:     uint32(testMeta.ModTime),
+			Time:     testMeta.ModTime,
 			Filename: testMeta.Filename,
 		},
 		keyRingTestPrivate,
